Moved extension methods to right place
